+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Enh #5: Refactor codebase to improve performance (@terabytesoftw)
 - Enh #6: Add `SvgBlock` enum for HTML SVG tag representation and implement `BlockInterface` class (@terabytesoftw)
 - Enh #7: Add `BaseSvgTag` and `Defs` classes for SVG block-level elements and related tests (@terabytesoftw)
+- Enh #8: Add `HasFill` trait and corresponding tests for managing SVG `fill` attribute (@terabytesoftw)
 
 ## 0.2.0 March 31, 2024